Remove the useless _split in OpenAPIv3 validator (unsure on this one)
_split() would split on the last value of the reference if an array ref was
passed in, but I can't find any instances where it was.

I can't find any documentation as to why this might be needed, even combing
git history. Further, I remove this and run the full test suite and things
run just fine. That's even with the extra env variables I need to turn
on to get all tests to run.